backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "total-llm"
version = "1.0.0"
description = "LLM-based Security Monitoring Platform"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.7.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "asyncpg>=0.29.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "cryptography>=41.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiohttp>=3.9.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "markdown>=3.5.0",
    "rank-bm25>=0.2.2",
    "reportlab>=4.0.0",
    "paramiko>=3.4.0",
    "requests>=2.31.0",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
]
kafka = [
    "kafka-python>=2.0.0",
    "aiokafka>=0.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/total_llm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
